requirements.txt
aiogram
aiohttp
pycryptodome
mysql-connector-python